refactored and split pointer_inference::state tests, refactored stack_substitution tests, fixed bug for expr! macro supporting negative constants.
